requirements.txt
# E-Bonding Platform — Core Dependencies
# Pin majors, allow patches for security updates.

# Web framework
fastapi>=0.110,<1.0
uvicorn[standard]>=0.29,<1.0
pydantic>=2.7,<3.0

# Database
sqlalchemy>=2.0,<3.0
asyncpg>=0.29,<1.0           # PostgreSQL async driver
alembic>=1.13,<2.0           # Migrations

# Vector search (hybrid discovery — FR-03)
pgvector>=0.2,<1.0           # pgvector SQLAlchemy integration

# PDF generation (document assembly — FR-05)
weasyprint>=62,<63           # PDF/A-2b generation
Jinja2>=3.1,<4.0             # Template engine for bond documents

# Cryptography (HSM signing, manifest validation — FR-06)
cryptography>=42,<44
PyJWT>=2.8,<3.0              # JWT auth

# Object storage
boto3>=1.34,<2.0             # S3 for audit bundles, RON recordings

# JSON Schema validation
jsonschema>=4.22,<5.0

# HTTP client (RON adapter, KYC provider, ledger client)
httpx>=0.27,<1.0

# Observability
structlog>=24.1,<25.0        # Structured logging
opentelemetry-api>=1.24,<2.0
opentelemetry-sdk>=1.24,<2.0

# Task queue (async rule evaluation, bundle generation)
celery[redis]>=5.3,<6.0

# Configuration
python-dotenv>=1.0,<2.0
